get port stats
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@33314 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/linux/universal/linux-3.18/include/linux/switch.h
@@ -59,8 +59,8 @@ struct switch_port_link {
 };
 
 struct switch_port_stats {
-	unsigned long tx_bytes;
-	unsigned long rx_bytes;
+	unsigned long long tx_bytes;
+	unsigned long long rx_bytes;
 };
 
 /**
@@ -115,12 +115,12 @@ struct switch_dev {
 	const char *alias;
 	struct net_device *netdev;
 
-	int ports;
-	int vlans;
-	int cpu_port;
+	unsigned int ports;
+	unsigned int vlans;
+	unsigned int cpu_port;
 
 	/* the following fields are internal for swconfig */
-	int id;
+	unsigned int id;
 	struct list_head dev_list;
 	unsigned long def_global, def_port, def_vlan;
 
@@ -148,8 +148,8 @@ struct switch_portmap {
 
 struct switch_val {
 	const struct switch_attr *attr;
-	int port_vlan;
-	int len;
+	unsigned int port_vlan;
+	unsigned int len;
 	union {
 		const char *s;
 		u32 i;
